create scoped v1 namespace, move views and controllers
