+++ sophia/app/controllers/v1/base_controller.rb
@@ -1,0 +1,5 @@
+module V1
+  class BaseController < ApplicationController
+
+  end
+end
