--- sophia/app/views/users/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @users, partial: "users/user", as: :user
