--- sophia/app/views/users/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! "users/user", user: @user
