+++ sophia/app/views/v1/users/_user.json.jbuilder
@@ -1,2 +1,2 @@
 json.extract! user, :id, :created_at, :updated_at
-json.url user_url(user, format: :json)
+json.url v1_user_url(user, format: :json)
