+++ sophia/app/views/v1/users/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.array! @users, partial: "v1/users/user", as: :user
