+++ sophia/app/views/v1/users/show.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! "v1/users/user", user: @user
